extract common tableidgenerator stuff
