+++ airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/integrations/destination/s3_data_lake/TableIdGenerator.kt
@@ -5,12 +5,6 @@
 package io.airbyte.integrations.destination.s3_data_lake
 
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.command.iceberg.parquet.GlueCatalogConfiguration
-import io.airbyte.cdk.load.command.iceberg.parquet.NessieCatalogConfiguration
-import io.airbyte.cdk.load.command.iceberg.parquet.RestCatalogConfiguration
-import io.airbyte.cdk.load.data.Transformations
-import io.micronaut.context.annotation.Factory
-import javax.inject.Singleton
 import org.apache.iceberg.catalog.Namespace
 import org.apache.iceberg.catalog.TableIdentifier
 
@@ -30,44 +24,6 @@ class SimpleTableIdGenerator(private val configNamespace: String? = "") : TableI
     }
 }
 
-/** AWS Glue requires lowercase database+table names. */
-class GlueTableIdGenerator(private val databaseName: String?) : TableIdGenerator {
-    override fun toTableIdentifier(stream: DestinationStream.Descriptor): TableIdentifier {
-        val namespace =
-            Transformations.toAlphanumericAndUnderscore(
-                (stream.namespace ?: databaseName)!!.lowercase()
-            )
-        val name = Transformations.toAlphanumericAndUnderscore(stream.name.lowercase())
-        return tableIdOf(namespace, name)
-    }
-}
-
-@Factory
-class TableIdGeneratorFactory(private val s3DataLakeConfiguration: S3DataLakeConfiguration) {
-    @Singleton
-    fun create() =
-        when (s3DataLakeConfiguration.icebergCatalogConfiguration.catalogConfiguration) {
-            is GlueCatalogConfiguration ->
-                GlueTableIdGenerator(
-                    (s3DataLakeConfiguration.icebergCatalogConfiguration.catalogConfiguration
-                            as GlueCatalogConfiguration)
-                        .databaseName
-                )
-            is NessieCatalogConfiguration ->
-                SimpleTableIdGenerator(
-                    (s3DataLakeConfiguration.icebergCatalogConfiguration.catalogConfiguration
-                            as NessieCatalogConfiguration)
-                        .namespace
-                )
-            is RestCatalogConfiguration ->
-                SimpleTableIdGenerator(
-                    (s3DataLakeConfiguration.icebergCatalogConfiguration.catalogConfiguration
-                            as RestCatalogConfiguration)
-                        .namespace
-                )
-        }
-}
-
 // iceberg namespace+name must both be nonnull.
-private fun tableIdOf(namespace: String, name: String) =
+fun tableIdOf(namespace: String, name: String): TableIdentifier =
     TableIdentifier.of(Namespace.of(namespace), name)
